requirements.txt
bs4==0.0.1
html-table-parser==0.1.0
july==0.1.3
matplotlib==3.5.3
numba==0.57.1
numpy==1.21.6
packaging==23.1
pandas==1.3.5
patsy==0.5.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytimekr==0.1.0
pytz==2023.3
pywin32==306
requests==2.30.0
scipy==1.10.1
selenium==4.14.0
six==1.16.0
statsmodels==0.14.0
tslearn==0.6.1
tzdata==2023.3
yellowbrick==1.5